+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandDescribe.cs
@@ -4,6 +4,7 @@
 // RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
+using System.Linq;
 using System.Reflection;
 using System.Text;
 using MapsDirectlyToDatabaseTable;
@@ -40,6 +41,13 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
 
             foreach (PropertyInfo p in o.GetType().GetProperties())
             {
+                // don't describe helper properties
+                if(p.GetCustomAttributes(typeof(NoMappingToDatabase)).Any())
+                {
+                    continue;
+                }
+
+
                 sb.AppendLine(p.Name + ":" + (p.GetValue(o)?.ToString() ?? "NULL"));
             }